fix static test data creation
renamed InformationSchemaQueryTest to
StaticInformationSchemaQueryTes and fixed the optimization
to only create the tables once for the class.

It didn't really work because Test classes are instantiated
for each test method.
